+++ miyoo.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#if defined(MIYOOMINI)
+
+#include "configuration.h"
+
+void miyoo_event_fullscreen_impl(settings_t* settings);
+
+#endif
