+++ retroarch.c
@@ -211,6 +211,10 @@
 #include "lakka.h"
 #endif
 
+#if defined(MIYOOMINI)
+#include "miyoo.h"
+#endif
+
 #define _PSUPP(var, name, desc) printf("  %s:\n\t\t%s: %s\n", name, desc, var ? "yes" : "no")
 
 #define FAIL_CPU(simd_type) do { \
@@ -2085,29 +2089,6 @@ bool is_accessibility_enabled(bool accessibility_enable, bool accessibility_enab
 }
 #endif
 
-#if defined(MIYOOMINI)
-static inline void show_miyoo_fullscreen_notification(settings_t* settings){
-   char msg[80];
-   struct retro_message_ext msg_obj = {0};
-
-   msg[0] = '\0';
-
-   snprintf(msg, sizeof(msg), "Aspect: %s. %s.",
-      settings->bools.video_dingux_ipu_keep_aspect ? "Original" : "4:3",
-      settings->bools.video_scale_integer ? "1x Scale" : "Screen Fill");
-
-   msg_obj.msg      = msg;
-   msg_obj.duration = 1000;
-   msg_obj.priority = 3;
-   msg_obj.level    = RETRO_LOG_INFO;
-   msg_obj.target   = RETRO_MESSAGE_TARGET_ALL;
-   msg_obj.type     = RETRO_MESSAGE_TYPE_STATUS;
-   msg_obj.progress = -1;
-
-   runloop_environment_cb(RETRO_ENVIRONMENT_SET_MESSAGE_EXT, &msg_obj);
-}
-#endif
-
 /**
  * command_event:
  * @cmd                  : Event command index.
@@ -3781,13 +3762,7 @@ bool command_event(enum event_command cmd, void *data)
 #endif
       case CMD_EVENT_FULLSCREEN_TOGGLE:
 #if defined(MIYOOMINI)
-         settings->bools.video_dingux_ipu_keep_aspect = !settings->bools.video_dingux_ipu_keep_aspect;
-         if (!settings->bools.video_dingux_ipu_keep_aspect) {
-            settings->bools.video_scale_integer = !settings->bools.video_scale_integer;
-         }
-         video_driver_apply_state_changes();
-
-         show_miyoo_fullscreen_notification(settings);
+         miyoo_event_fullscreen_impl(settings);
 #else
          {
             audio_driver_state_t
